Don't include the `innodb' suite in the list of default test suites.
This suite is for the innodb plugin, which we do not use in MariaDB
(we instead have XtraDB as default storage engine, which includes the
innodb plugin).

We have included the same tests in the main suite, so no point in
running them twice (and most of them are disabled anyway due to the
absense of the innodb plugin).