Fix Windows test failures in binlog tests in certain time zones.
On Unix systems TZ can be set to change the time zone for specific processes only.
But on Windows this does not fully work. It changes some aspects of time zones
in the system but not others (notably localtime() vs. file system time stamps).

This causes test failures in Windows in certain time zones (but not all),
where PURGE BEFORE DATE statements get the wrong files when TZ is set to
+03:00 by default.

Fix by only setting TZ in the small number of tests that really need it, and leave
it untouched in the rest.

Thanks to Alex Budovski for helping with this.